backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
feedparser==6.0.10
requests==2.32.3
Pillow==10.1.0
python-dotenv==1.0.0
pytrends==4.9.2
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
yt-dlp==2024.3.10
